more clean up and untangling
